tweak cct schema to add name & make path optional
We have a (currently undocumented) feature where a CCT module can
be defined in terms of a name only, and if a directory matching that
name exists in a 'cct' directory adjacent to the image.yaml we are
processing, it is used as the module source.

This feature did not work since the schema was introduced because
the schema defined the 'path' key as mandatory. This change relaxes
that and also defines the optional 'name' key that this feature
relies upon.