move EXPOSE instructions towards end of template
EXPOSE instructions do not take effect during build time, so it doesn't
matter where they are in the Dockerfile. However, putting them early in
the template means that if the user changes ports, all instructions
after EXPOSE are invalided in Docker's build cache.

Move it towards the end so a user can change ports in their image.yaml,
regenerate to an existing build directory and rebuild without
invalidating (possibly expensive) instructions.

We could do the same for LABELs, but not ENVs (which might be depended
upon by scripts run at build-time), and it's probably a good idea to
keep the LABELs and ENVs together.